Install jasmine
First bit of project setup. We're going to need a JS test framework that
runs on Node and the browser. Jasmine is small so less of a hassle to
fetch over a bad connection. It also has built-in mocking support which
will probably be useful for doing multiple storage backends.