Collate ingredients across recipes
This is mostly for assembling shopping lists, so we collate
ingredients and add up the quantities if we can. We also collate
ingredients by "group" (such as supermarket aisle or shopkeeper)
as an aide during a shopping trip.

Ingredients without a group are placed in an 'undefined' group at
the start of the groups. Other groups, and ingredients within a
group, are sorted lexically on name.