Put test data in the 'spec' directory
We have been running tests off the example recipes in dr.syntax,
primarily so that the code and tests are based off real data. However,
when these are added to, it breaks the tests, so we're now taking a
snapshot of what we've collected so far and putting in the 'spec'
directory, where it's stable.