Extract filesystem adapter from Store
To allow recipe data to be accessed from the browser, we want to break
the coupling between Store and Node's fs module. We currently only need
read access to files, and this can be implemented on top of fetch() in
the browser.

To move towards this, here we extract the fs-specific code from Store
into the FSAdapter class, and pass that into Store as a dependency. As
fetch() cannot list directories or call stat(), we've switched to
requiring an explicit recipe index file.