Implement search in the React app
This uses the chrome design from the mocks, but the HTML structure
has changed slightly so the shopping list mock will no longer work
fully. (That will disappear completely when the shopping list is
implemented in React.)

web/recipes needs to contain the recipes (including the recipe
index recipes.json, and ontology.json) -- for now we are doing
that by symlinking to dr.syntax, but that will likely change to
a more fullsome build step at some point.