Don't load recipes for the shopping list
We were (async) loading some recipes to populate the shopping list
component. Shopping lists now start with no recipes, so there's no
point in doing this any more.

This means that the shopping list component is now useless until
we have a way of adding recipes to the underlying shopping list
object.