Add a popup to adjust recipe counts
In the shopping list, we have a number of recipes. If you tap one,
you'll get a popup which allows you to increase or decrease the
count of that recipe in the shopping list, with the ingredients
required (and their quantities) being recalculated as you do. If
you set the count to 0 it will remove the recipe completely.