SLIA knows if it satisfies SLI
SLIA indicates "what we have bought" (or is in our cupboard or
similar). SLI is "what we need". SLIAs keep some state which helps
them determine they can satisfy an SLI even when there are multiple
units in play. (Although it won't always get it right, but the
failure mode here is that the user has to quickly check.)