Begin sync adapter
To work for real in the browser, we'd like an implementation of
read_file() such that Store can load content from a user's Dropbox. This
should work offline by caching items in localStorage.

Here we begin this work. read_file() inspects the cache and returns a
result immediately if one exists. Then it tries to refresh the cache
from the remote server by calling the Dropbox client. If it gets a
response, it updates the cache.

This strategy means the caller will sometimes be given content that's
stale, but this is a better offline or high-latency UX than blocking on
the request to Dropbox.