add() and update() for the Dropbox client
The /files/upload method of the Dropbox API implements optimistic
concurrency control: all files have a version, which Dropbox calls the
'rev'. It returns the rev with the metadata when you download a file,
and returns the new one when you successfully upload one. To upload a
new version of a file, you must specify whether you believe you are
adding a new file, or modifying an existing one. If the latter, you must
supply the file's current rev for your upload to succeed.

Here we implement two methods as a wrapper over this functionality.
add() takes a file path and data, and update() takes these plus a rev
value.

On success, these methods return a Reponse whose meta field contains the
new rev. On conflict errors, a special type of error is thrown so the
caller can detect it. All other kinds of errors result in a generic
exception.