Implement write_file() for the sync adapter
Now that the Dropbox client has an interface for updating files, we can
use it to store content from the browser. However we need to do this in
a way that won't lose updates if the application makes multiple writes
close together, if the client goes offline, or even if the user closes
the app and returns to it.

To support this, the Cache class (backed by a Storage object, usually
localStorage), gains a new method called write(). Data written by the
application is stored here initially, and the record is marked as dirty
using an atomic counter. The set of dirty records is persisted so it can
be recalled on page reloaded. Having written the data to the cache,
SyncAdapter calls Sync.process() to trigger a sync attempt for the
record.

The Sync class handles synchronising the cache with a "remote", an
abstraction implemented by the Dropbox client. Its refresh() method
retrieves a document from the remote and stores it in the cache (along
with its metadata like its 'rev'). This fails silently, so that the
cache just contains whatever we last managed to successfully fetch from
the remote. This request is retried whenever read_file() is called on
the adapter, to attempt to update the cache.

The process() method handles writing in the other direction, from the
cache to the remote. This requests a copy of the cache-record, which
includes the record's data, its 'rev' (Dropbox's name for the version
ID), and the current value of the dirty-counter, so we know which local
version we're sending.

It attempts to write this record to the remote. If this fails due to a
conflict, then we re-fetch the record so we have its new 'rev' value,
and schedules a retry. If it succeeds then we call Cache.resolve() with
the cache-record and the remote response. Cache.resolve() stores the new
'rev' and returns true if the record is now up-to-date, meaning its
dirty-counter has not changed during this sync attempt. If this is not
the case, we schedule another sync attempt as new data has been written
while we were syncing.

For testing purposes, SyncAdapter has a method called wait_sync() that
blocks until all pending dirty records have been synchronised.

To make this production-ready I'd want a few further changes.

First, this does not implement merging. If a record is dirty, then
fetching it from the remote does not clobber the currently cached data.
When we sync that data, the local copy will overwrite whatever is in the
remote. This means anything you write in the current session should be
saved, which is better than nothing, but you'd want to incorporate
merging to make this really robust. You'd also need the storage layer to
inform the UI that new data has arrived from the remote.

Second, I'd want the Sync class to be smarter about scheduling.
Currently, it schedules retries by waiting one second and then trying
again, with a flag to prevent multiple concurrent sync attempts for the
same record. You'd probably want to use a queue for this instead, and
implement something that backs off on making requests if we detect that
they keep failing, indicating we're offline. It should also flush the
current dirty records when the app first boots.