Add an about page
Similar text to the README. Make this the default screen, which may be
the wrong choice but feels slightly more sensible when you load up the
very first time. (It'd certainly make sense to change this if you have
a shopping list -- so if anyone integrates the Dropbox backend, we can
have a look at that.